chore: correct Credfeto.ChangeLog.BenchMark.Tests casing to Benchmark.Tests
Renamed the project directory, csproj, and namespaces from
Credfeto.ChangeLog.BenchMark.Tests to Credfeto.ChangeLog.Benchmark.Tests,
correcting a "BenchMark" casing typo, and updated the solution file and
GitHub label/labeler config that referenced the old name.

Also removes XunitAutoGeneratedEntryPoint=false from the renamed csproj,
fixing the pre-existing CS5001 "no entry point" failure in the
Release-config solution build (#372) - the pre-commit buildtest hook
cannot pass without this fix present, so it is included here rather
than as an unbuildable intermediate commit.
