+++ .github/labeler.yml
@@ -15,7 +15,7 @@
 "credfeto-changelog":
   - any: ['src/Credfeto.ChangeLog/**/*']
 "credfeto-changelog-benchmark-tests":
-  - any: ['src/Credfeto.ChangeLog.BenchMark.Tests/**/*']
+  - any: ['src/Credfeto.ChangeLog.Benchmark.Tests/**/*']
 "credfeto-changelog-cmd":
   - any: ['src/Credfeto.ChangeLog.Cmd/**/*']
 "credfeto-changelog-cmd-tests":
