+++ src/Credfeto.ChangeLog.Benchmark.Tests/Credfeto.ChangeLog.Benchmark.Tests.csproj
@@ -39,7 +39,6 @@
     <UseSystemResourceKeys>true</UseSystemResourceKeys>
     <WarnOnPackingNonPackableProject>false</WarnOnPackingNonPackableProject>
     <WarningsAsErrors />
-    <XunitAutoGeneratedEntryPoint>false</XunitAutoGeneratedEntryPoint>
   </PropertyGroup>
   <ItemGroup>
     <!-- error NU1903: Warning As Error: Package 'System.Private.Uri' 4.3.0 has a known high severity vulnerability -->
